Exceptions and MigrateDatabaseAsync extension method were moved to the newly created Hotel.Shared assembly
