+++ Booking/Booking.Api/Program.cs
@@ -1,4 +1,5 @@
-using Booking.Infrastructure;
+using Booking.Infrastructure.Database;
+using Hotel.Shared.Extensions;
 
 namespace Booking.Api;
 
@@ -7,7 +8,7 @@ public class Program
     public static async Task Main(string[] args)
     {
         var host = CreateHostBuilder(args).Build();
-        await host.MigrateDatabaseAsync();
+        await host.MigrateDatabaseAsync<ApplicationContext>();
         
         await host.RunAsync();
     }
