--- Booking/Booking.Domain/Exceptions/BadRequestException.cs
@@ -1,5 +1,0 @@
-using System.Net;
-
-namespace Booking.Domain.Exceptions;
-
-public class BadRequestException(string message) : RequestException(HttpStatusCode.BadRequest, message);
