--- Booking/Booking.Domain/Exceptions/NotFoundException.cs
@@ -1,5 +1,0 @@
-using System.Net;
-
-namespace Booking.Domain.Exceptions;
-
-public class NotFoundException(string message) : RequestException(HttpStatusCode.NotFound, message);
