--- Booking/Booking.Domain/Exceptions/RequestException.cs
@@ -1,8 +1,0 @@
-using System.Net;
-
-namespace Booking.Domain.Exceptions;
-
-public class RequestException(HttpStatusCode statusCode, string message) : Exception(message)
-{
-    public HttpStatusCode StatusCode = statusCode;
-}
