--- Booking/Booking.Infrastructure/HostExtensions.cs
@@ -1,22 +1,0 @@
-using Booking.Infrastructure.Database;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace Booking.Infrastructure;
-
-public static class HostExtensions
-{
-    public static async Task<IHost> MigrateDatabaseAsync(this IHost host)
-    {
-        using (var scope = host.Services.CreateScope())
-        {
-            var services = scope.ServiceProvider;
-
-            var context = services.GetRequiredService<ApplicationContext>();
-            await context.Database.MigrateAsync();
-        }
-
-        return host;
-    }
-}
