+++ Hotel.Shared/Extensions/HostExtensions.cs
@@ -1,19 +1,18 @@
-using Managing.Infrastructure.Database;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Managing.Infrastructure;
+namespace Hotel.Shared.Extensions;
 
 public static class HostExtensions
 {
-    public static async Task<IHost> MigrateDatabaseAsync(this IHost host)
+    public static async Task<IHost> MigrateDatabaseAsync<T>(this IHost host) where T: DbContext
     {
         using (var scope = host.Services.CreateScope())
         {
             var services = scope.ServiceProvider;
 
-            var context = services.GetRequiredService<ApplicationContext>();
+            var context = services.GetRequiredService<T>();
             await context.Database.MigrateAsync();
         }
