+++ Identity/Identity.Api/Program.cs
@@ -1,4 +1,6 @@
+using Hotel.Shared.Extensions;
 using Identity.Infrastructure;
+using Identity.Infrastructure.Database;
 
 namespace Identity.Api;
 
@@ -7,7 +9,7 @@ public class Program
     public static async Task Main(string[] args)
     {
         var host = CreateHostBuilder(args).Build();
-        await host.MigrateDatabaseAsync();
+        await host.MigrateDatabaseAsync<ApplicationContext>();
         await host.SeedDataAsync();
         
         await host.RunAsync();
