--- Identity/Identity.Domain/Exceptions/NotFoundException.cs
@@ -1,5 +1,0 @@
-using System.Net;
-
-namespace Identity.Domain.Exceptions;
-
-public class NotFoundException(string message) : RequestException(HttpStatusCode.NotFound, message);
