--- Identity/Identity.Domain/Exceptions/RequestException.cs
@@ -1,8 +1,0 @@
-using System.Net;
-
-namespace Identity.Domain.Exceptions;
-
-public class RequestException(HttpStatusCode statusCode, string message) : Exception(message)
-{
-    public HttpStatusCode StatusCode { get; set; } = statusCode;
-}
