+++ Identity/Identity.Domain/Identity.Domain.csproj
@@ -13,4 +13,8 @@
       <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\..\Hotel.Shared\Hotel.Shared.csproj" />
+    </ItemGroup>
+
 </Project>
