+++ Identity/Identity.Infrastructure/HostExtensions.cs
@@ -20,17 +20,4 @@ public static class HostExtensions
 
         return host;
     }
-
-    public static async Task<IHost> MigrateDatabaseAsync(this IHost host)
-    {
-        using (var scope = host.Services.CreateScope())
-        {
-            var services = scope.ServiceProvider;
-
-            var context = services.GetRequiredService<ApplicationContext>();
-            await context.Database.MigrateAsync();
-        }
-
-        return host;
-    }
 }
