+++ Managing/Managing.Api/Program.cs
@@ -1,4 +1,5 @@
-using Managing.Infrastructure;
+using Hotel.Shared.Extensions;
+using Managing.Infrastructure.Database;
 
 namespace Managing.Api;
 
@@ -7,7 +8,7 @@ public class Program
     public static async Task Main(string[] args)
     {
         var host = CreateHostBuilder(args).Build();
-        await host.MigrateDatabaseAsync();
+        await host.MigrateDatabaseAsync<ApplicationContext>();
         
         await host.RunAsync();
     }
