+++ Managing/Managing.Domain/Managing.Domain.csproj
@@ -11,4 +11,8 @@
       <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\..\Hotel.Shared\Hotel.Shared.csproj" />
+    </ItemGroup>
+
 </Project>
